moved the retrieveRecipients function from the controller logic to the helper function
